FES_Interface/requirements.txt
Pillow==9.5.0
Pillow==10.3.0
pyserial==3.5
